Player Development Day 2
Fixed player movement in the direction by changing the X axis of the camera when the mouse moves rather than the player.
Added in functionality for sprinting and crouching when the left shift or x keys are pressed (unable to use control in the unity editor).

---TO DO STILL---
Fix movement where if a horizontal key and vertical movement key are pressed at the same time they do not have more speed.