Player Development  Day 3
Added base weapon model as a rectangle. Placed it under camera hierachy and attatched a bullet spawn to it.
Created a base form of projectile to spawn as well.
If player runs out of ammo they cannot fire anymore.

--TO DO STILL--
Fix movement issue in previous update, make basic UI for player functionality so far.