Player Development Day 5
I have a good understanding of canvas now.
Created UI positions for player health, weapon name, the amount of ammo left, and the timer of the game.

Next time I will add more onto the UI setting up defaults for the rest of my design as well as add borders to the text on the UI.