Environment Development Day 3
Fixed the maps depth adding in ledges and a upper walkway for the right area. Messed with positions of slanted planes.
Added in planes on the opposite side of current planes to show the bottom as well.
Fixed crouching issue and rewrote player code. Added in funtionality to prevent player from hopping down a slope and they are now grounded as normal.
Started to add item positions based on empty gameobjects as well as the temporary prefabs for the items.

---TO DO NEXT----
Spawn ammo and health crate/boxes dynamically and put walls up for the rooms and hallways.