Environment Development Day 4
Started Adding Walls to the planes in the map. The walls are Cubes resized with box colliders.
I also switched all the flat planes from mesh colliders to box colliders to be more efficient.
I created a script for spawning items for the ammo and health boxes.
Currently as it stands it will spawn either a ammo or a health item on a 1 or a 0. The problem with this is that they all could be ammo items or health items at the worst case.
Currently every 20 seconds they will be despawned and a new set will be populated each time.
This is a deviation to the original design as if an item is not taken by a player it will remain. There is a potential then for all of them to become health items when a player may not need them causing the player to have to take damage to clear up a space.
This should hopefully fix that issue for now.
The player can also walk into the items and destroy them regaining health and ammo respectively.
Another thing I noticed is the main rooms seem a bit bigger than expected however that is mainly due to them being emptier as well. This could also allow the player to face more enemies and in the future can be resized down if the space is not needed.

I will continue finishing the walls next time.