Environment Development Day 5
Finished making the walls the rooms and hallways. Some are a bit sloppy and will need tidying. I just realized what the proportions were for the axis's so when I edit them it will be a lot faster as well.
So far to do a lap in a circle of the map it takes around 80 seconds.
Bit longer than I originally expected but I can scale down if needed.

I will return to tweak the walls and work on the bullet collisions with the environment after.