Environment Development Day 6
I tweaked the wall heights today to not be as tall and imposing.
I also scaled down the overall map so a lap while running can be made in 60 seconds.
Finally used a friends bullet model that he created as the new projectile for the main weapon and created a script that allows the bullet to be destroyed when it collides with another object.
I will add a bullethole texture if it collides on a part of the map, as well as a sound and more effects later on.