Enemy Development Day 1
Originally when implementing the enemy design I planned on having the spawns randomly generated away from the player without fixed positions. This proved to be rather complicated as well as time consuming. It also could be potentially more resource consuming depending on the points it tried to place them at.
I also struggled when importing the zombie model initially getting it to work and function. Currently I have the animation set to Idle, which came with the model. This is of course better than the base t-pose.
I am having trouble working a generic collider with it for now getting my bullets to collide with it 100% of the time and destroying it. It works most of the time however atm but I will need to revisit this.
I have implemented base features tied with the score system so far updating the enemy count on the UI, the wave number as well as setting the players score based on how many enemies are killed too.
I will have to continue working on this a bit more and soon get into the main problem of developing the AI once the foundation for their spawns and the player's interactions with the enemies are more developed.