Enemy Development Day 2
Originally I planned on doing a grid over  the map and designed a lot of the game to use said grid. Pathfinding for the enemy was to be one of the elements that utilized the grid. After a few days of research and trying to work a grid properly on my map that I had manually created, it ended up being more feasible  to manually create the pathfinding routes rather than using a grid for the manually created map.

This ended up with me creating Waypoint nodes that would be used to direct the enemy agents and keep them on set path allowing them to reach different areas. The main goal of using the A* pathfinding way is still used as multiple classes were developed to allow this including, NodeController, PriorityQueue and PathFinder. I originally tried to copy and reproduce the examples potrayed in one of the books i used for research, however they way they did the grid I was unable to reproduce which led to a lot of confusion for me. In the end the current pathfinding is similar in execution to the way the book's example is now, however with manually created Nodes with neighbors I have manually assigned.

I will try to adapt this to the enemy zombie model next to have it patrol around the map. If this works I can start work on it's behavior tree.