Enemy Development Day 3
During this period, I had trouble understand how to actually implement behaviour trees. It took me more research and trial and error as I got the zombies to rotate and move towards the initial node of their path but they would not move afterwards to the rest. After a lot of trial and error and research I found https://github.com/GrymmyD/UnityBehaviourTree which contained project code and information I found invaluable. The previous books I used were helpful but too confusing to replicate the same example as it was too small case and not as helpful. I have replicated the logic for the composite nodes and the framework but all the leaf nodes and core gameplay logic is all mine. I now have the zombies able to patrol around the map going to different nodes as well as swap between walking and idle animations. The next step will to be to have them fall down before they die and despawn.