Enemy Development Day 5
Added in behaviors for following the player on sight.
Also added in sight for the zombies based on the example code from the programming book 3rd edition.
Also added in checks to see if the target is visible or not and got the main selector of the behavior tree working. Added in the framework for checking if the player is in melee range. I will have to add some functionality howeverr for if the enemy gets stuck walking into a wall in rare cases and if the player is on a higher level. I will have to look at planning a path to get to the player rather than moving to the players current position.