Enemy Development Day 6
Developed behaviour tree for melee functionaltity. Had an issue for a while where the tree logic was getting confused between chasing and meleeing. This cause the animations to trigger back and forth repeatdly. It turned out the melee animation would sometimes push a bit back causing it to enter the previous sequence and repeatdly switch animated states. This was fixed.
I need to add functionality still for moving when the player is on another level. I also need to add colliders or find a way for a collider to be registered to the zombie's hands. I am having trouble with that atm as well as the colliders do not follow the zombies animations at this time and its transform.