Enemy Development Day 6 Part 2
Added in colliders for the zombies attacks. It turned out i had to attatch it to the bone ligments for the colliders to move with the animation.
The player's health can now decrease.
When it reaches 0 the player is presented with a game over screen.
The next step is to fix the AI''s movements when chasing if the player is on another level. Add in the actual score to the game over screen as well.
Then I can start working on the different colored versions.