Enemy Development Day 7
Added in different color types for the zombie enemy type.
Added gravity as well to the zombies so they fall rather than glide when walking off a platform.
Added in functionality for the score modifier which takes note on when the player takes damage. Also adjusted some more score data and information.
Having a problem expanding the pathing though for when the player is on another plane/cube/elevation to the AI. I tried at first to get a path if the player's Y position was a lot higher but that threw alot of issues. I am trying to add in a way to get it to get a path to the player until they are both on the same plane however I am currently having quite a bit of issues with it and I have been struggling recently this past week. I am hoping to resolve this by before the end of the week and start adding in another enemy type.