Enemy Development Day 8
I am still having problems with the enemy moving to the player when on a different plane. I got it working somewhat but I was presented with a large amount of bugs and errors as well. When I return I will create a dictionary of all the pathnodes associated with the plane they are on top of. Then I will get the nearest node to the plane the player is on as well as for the enemy.