Enemy Development (Mainly Pathing) Day 9
Added more nodes in for the pathing to keep it more accurate going from 52 total to over 120 to add more routes and keep the AI more controlled.
Originally was updating the AI path to the player if they were on two different planes based on the end node the player was by.
It worked decently then apart from backtracking to a certain point at times whenever the new path was created. I will try to do it for if the players different plane is different from the previous one.