Enemy Development Day 9 part 2
Fixed plane issue as well as if the enemy has no line of sight after reaching chase destination for path. Also added in function that checks if node is behind it so it doesnt double back to go to it on the path. Downloaded and imported soldier pack to try and start working on that this weekend.