Enemy Development Day 10
Ok at this point I have created a soldier behaviour tree, as well as actions and tested the soldiers tree. I had trouble with the shooting for a while making it hit the player if it crouched but that is fixed now directing the bullet towards the player instead of instantiating it where it spawns and going forward.
I also had trouble initally with the animations as the soldier has fewer and is more limited. I also had some issues with the ability to make the soldier walk backwards towards the previous closest node. This required alot of testing and required me to change the angle or direction for that instance allowing him to walk backwards. Sometimes there are some issues with him switching between animations too fast in the tree but I can come back to that later. I also changed the patrol movement tree and chase for the zombies and soldiers where they now go to a random offset on the X and Z axis rather than the exact position. This should help reduce clashes and collisions between AI enemies at different speeds. This has caused some issues with chasing the player at times providing rare instances where the AI needs to backpedal to a previous point to resume the chase.
Currently I will try to implement the soldiers into the waves now. I will also try to fix an issue where the player is in the soldiers collider causing it to do nothing. The only other issue aside from that is sometimes going down a certain slope the soldier will lose sight and walk off. The sight class will need to be looked at. Afterwards I can tidy up code and then start work on the sounds and hearing for the AI.