Enemy Development Day 10 Part 2
Created Blue, red and green version of soldiers. Added them to the wave spawner. Adjusted waves so the first wave is soldiers, the second wave is zombies and then every wave after is a random mixture. Tommorow will work on the Game over score screen and the sound perception as well as sounds for the AI and player. Also bug when reloading/restarting the game with the camera lower transform than it should.