Player Feedback Day 1 Part 2
Added in where if the player takes damage their screen flashes red similar to enemy. Essentially changed the opaquness of a panel to be brighter if this happens.

Also fixed error, where if the player is within a certain range it doesnt matter of line of sight as this isnt a stealth game. If the player is within a certain range of the soldier, like 5 the soldier will automatically know the player is there. Added in offset for zombies for when in melee range based on attacking as well as it can push itself back/sometimes be .01 away rather than just 2 or less.
This fixes the issue when they are near.
-------------------------------------------
Need to start working on the hearing perception for the enemies.
The enemies doubling back to the previous path nodes now that offsets have been applied to the patrolling and chase positions.
Also need to fix the issue when they go up and down the ramps.