Sound Development Day 1
Rescaled UI for 1920x1080p.
Added in sound gun shot audio with fade out to prevent popping noise when the clip ends. Code found from stackoverflow for fade coroutine.
Created Settings menu for main menu and pause menu that allows the user to adjust the sound volumes for the music and gun shot fx.

Fixed issue where player took damage from zombies on the ground as well as shots curving upwards.
Still need to work on players melee function as well as enemy's reaction to the sounds.
As well as previous unresolved bugs.