Player Development Update 1
Added in the ability for the player to melee the enemy using right click.
Extends the weapon and attatched a collider as well as a rigidbody.
Had problems at first especially when trying to get the rotation and position correct after the action.
Ended up solving the issue using the local values rather the globals of its parent object.