Enemy Senses Update Day 1
Updated enemies to give them a simple hearing solution.
It seemed infeasible and ineffcient to have them listen for audio.
Changed it to when player fires a shot any enemies within a radius of the player will be alerted and chase.
Also added in a god mode toggle.
Fixed some other issues with the cursor on the screen during gameplay.
Need to fix bullets going through players at times due to being too fast.
Need to fix players walking on bullets at times.
Need to fix previous 2 pathing issues.
Also need to add credits in on the main menu.