Bullet Collision Development and Menu Credits
Added credits showing links and tools i used to help make the game.
Found help eventually to solve the issue where the bullets would not collide with the player due to the player being too fast or them walking on top.
Ended up with the result of when the bullet moves it shoots a raycast and if the player or enemy is in the next frame it destroys itself and inflicts damage.
Need to polish music slider, credits, fix the pathing bugs, maybe add a minimap.